% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_seq_methods.R
\name{affected_allele_count}
\alias{affected_allele_count}
\title{Determine total number of alleles shared by affecteds in a family}
\usage{
affected_allele_count(ped_haps, hap_map, ped_file)
}
\arguments{
\item{ped_haps}{sparse matrix.  The familial haplotype data.}

\item{hap_map}{data frame. Mapping data: maps individuals in \code{ped_file} to the haplotypes in \code{ped_haps}.}

\item{ped_file}{data frame.  Pedigree file, a for single family; i.e. not an entire study of ped files.}
}
\value{
A list of sharing counts, in the same order as the SNVs in \code{ped_haps}.
}
\description{
Determine total number of alleles shared by affecteds in a family
}
\keyword{internal}
