% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_ped}
\alias{check_ped}
\title{Checks individual pedigrees for proper format.}
\usage{
check_ped(ped_file)
}
\arguments{
\item{ped_file}{data.frame The pedigree.}
}
\description{
Checks individual pedigrees for proper format.
}
\keyword{internal}
