% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_peds}
\alias{check_peds}
\title{Checks ped_files for expected info and format.}
\usage{
check_peds(ped_files)
}
\arguments{
\item{ped_files}{Data frame. Must match format of pedigree simulated by sim_RVped}
}
\description{
Checks ped_files for expected info and format.
}
\keyword{internal}
