% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_SNV_map}
\alias{check_SNV_map}
\title{Check SNV_map for possible issues}
\usage{
check_SNV_map(SNV_map)
}
\arguments{
\item{SNV_map}{Data.frame. Must contain three columns with: column 1: marker names, must be listed in the same order as in the founder genotype file, column 2: the chromosomal position of the marker, column 3: the position of the marker in cM.}
}
\description{
INTENDED FOR INTERNAL USE ONLY
}
\keyword{internal}
