% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_Functions.R
\name{chias_count_BC}
\alias{chias_count_BC}
\title{Determine number of chiasmata before centromere}
\usage{
chias_count_BC(chiasmata_pos, center_loc)
}
\arguments{
\item{chiasmata_pos}{The simulated chiasmata postions}

\item{center_loc}{The centromere locations}
}
\value{
The number of chiasmata before the centromere
}
\description{
For internal use.
}
\examples{
ex_chias <- sim_chiasmataPositions(chrom_map = data.frame(start = 47, stop = 198))
ex_chias
chias_count_BC(ex_chias, 78)
}
