% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_exons.R
\name{combine_exons}
\alias{combine_exons}
\title{Combine overlapping exons into a single observation}
\usage{
combine_exons(exon_data)
}
\arguments{
\item{exon_data}{data.frame. This data.frame must include named variables: 'chrom', a chromosome identifer; 'exonStart', the first position of the exon in base pairs; and 'exonEnd', the last position of the exon in base pairs.}
}
\value{
A matrix of exon unions, i.e. non-overlapping observations, by chromosome.
}
\description{
Combine overlapping exons into a single observation
}
\examples{
# create a data frame that contains the
# the variables: chrom, exonStart, and exonEnd
exDat <- data.frame(chrom     = c(1, 1, 1, 2, 2, 2),
                    exonStart = c(1, 2, 5, 1, 3, 3),
                    exonEnd   = c(3, 4, 7, 4, 5, 6))

exDat

# supply exDat to combine_exons to combine overlapping segments
# into a single observation.
combine_exons(exDat)

data(hg_exons)
head(hg_exons)

hg2 <- combine_exons(hg_exons)
# notice that the combine_exons function disregards all
# variables other than 'chrom', 'exonStart', and 'exonEnd'
head(hg2)

# However, since the exons in hg_exons have already been
# combined into non-overlapping segments, the relevant
# variables remain unchanged
all(hg_exons[, c(1:3)] == hg2)

}
