% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_StudySeqFunctions.R
\name{condition_haploDist}
\alias{condition_haploDist}
\title{Condition Haplotype distribution}
\usage{
condition_haploDist(Chaplo_dist, RV_marker, RV_status)
}
\arguments{
\item{Chaplo_dist}{haplotype distribution for a single chromosome}

\item{RV_marker}{familial RV marker}

\item{RV_status}{0 or 1, 1 if RV is inherited}
}
\value{
The conditioned haplotype distribution.
}
\description{
Condition Haplotype distribution
}
