% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_seq_methods.R
\name{count_affectedRV}
\alias{count_affectedRV}
\title{Summary function for object of class fam_study}
\usage{
count_affectedRV(fam_study)
}
\arguments{
\item{fam_study}{The object returned by sim_RVstudy}
}
\value{
Not sure yet..
}
\description{
Summary function for object of class fam_study
}
\examples{
library(SimRVSequences)

#load pedigree, haplotype, and mutation data
data(study_peds)
data(EXmuts)
data(EXhaps)

#create variable is_CRV in EXmuts to identify causal
#rare variants, from which to sample familial variants.
EXmuts$is_CRV = FALSE
EXmuts$is_CRV[c(2, 3, 12, 24)] = TRUE

#supply required inputs to the sim_RVstudy function
seqDat = sim_RVstudy(ped_files = study_peds,
                     SNV_map = EXmuts,
                     haplos = EXhaps)

#to count the number of SNVs shared by the disease-affected
#relatives in each pedigree, supply the output returned by
#sim_RVstudy to count_affectedRV
count_affectedRV(seqDat)


}
