% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_map_functions.R
\name{create_chrom_map}
\alias{create_chrom_map}
\title{Create chromosome map from marker map}
\usage{
create_chrom_map(marker_map)
}
\arguments{
\item{marker_map}{Data.frame. Must contain three columns with: column 1: marker names, must be listed in the same order as in the founder genotype file, column 2: the chromosomal position of the marker, column 3: the position of the marker in cM.}
}
\value{
a dataframe catalouging the start and stop positions, in base pairs, for each chromosome.  We use this information to determine what regions to simulate recombination over.
}
\description{
Create chromosome map from marker map
}
