% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlimFunctions.R
\name{create_haplotypeSet}
\alias{create_haplotypeSet}
\title{Create list of haplotype matrices}
\usage{
create_haplotypeSet(markMap, genoMat)
}
\arguments{
\item{markermap}{The marker map returned by remao}

\item{genomat}{The full sparse matrix of genotypes}
}
\value{
A list of haplotype matrices
}
\description{
Create list of haplotype matrices
}
