% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_Functions.R
\name{estimate_haploDist}
\alias{estimate_haploDist}
\title{Estimate haplotype distribution}
\usage{
estimate_haploDist(pop_haplos, marker_map)
}
\arguments{
\item{pop_haplos}{Data.frame.  A data frame containing the population haplotypes}

\item{marker_map}{Data.frame. Must contain three columns with: column 1: marker names, must be listed in the same order as in the founder genotype file, column 2: the chromosomal position of the marker, column 3: the position of the marker in cM.}
}
\value{
haplo_dist The haplotype distribution
}
\description{
Estimate haplotype distribution
}

