% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXhaps.R
\docType{data}
\name{EXhaps}
\alias{EXhaps}
\title{Example Haplotypes Dataset}
\format{A sparseMatrix (formal class dgCMatrix) with 20000 rows and 173 variables.  Each row is the haplotype for 1 diploid individual, each column represents a single SNV locus.  Each column in \code{EXhaps} is described by a row in \code{\link{EXmuts}}.}
\usage{
EXhaps
}
\description{
This data set contains the haplotypes for 10000 diploid individuals along segments of length 20,000 base pairs from human chromosomes 1-5. This dataset is intended to accompany the EXmut dataset and is used for demonstration purposes.
}
\seealso{
\code{\link{EXmuts}}, \code{\link{read_slim}}
}
\keyword{datasets}
