% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_slimOut.R
\name{extract_tempIDs}
\alias{extract_tempIDs}
\title{Compile x,y pos of mutations in sparse matrix}
\usage{
extract_tempIDs(mutString, rarePos)
}
\arguments{
\item{mutString}{character. String containing mutations}

\item{rarePos}{numeric vector. position of variations}

\item{indPos}{numeric. row number of individual}
}
\value{
data.frame with x and y positions of mutations
}
\description{
Compile x,y pos of mutations in sparse matrix
}
\keyword{internal}
