% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_seq_methods.R
\name{famStudy}
\alias{famStudy}
\title{Constructor function for an object of class famStudy}
\usage{
famStudy(study_data)
}
\arguments{
\item{study_data}{The list of items returned by the \code{sim_RVstudy} function.}
}
\value{
an object of class \code{famStudy}.
}
\description{
Constructor function for an object of class famStudy
}
\keyword{internal}
