% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_StudySeqFunctions.R
\name{find_no_cSNV_rows}
\alias{find_no_cSNV_rows}
\title{Find haplotypes that do not carry any cRVs}
\usage{
find_no_cSNV_rows(haplos, RV_pool_loc)
}
\arguments{
\item{haplos}{sparseMatrix.  The sparseMatrix of genomes returned by \code{read_slimOut}.}

\item{RV_pool_loc}{The column locations of each SNV (in haplos)  for SNVs in the pool of cRVs.}
}
\value{
The reduced haplotype matrix.
}
\description{
Find haplotypes that do not carry any cRVs
}
\keyword{internal}
