% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_Functions.R
\name{gam_order}
\alias{gam_order}
\title{Assign gamete order}
\usage{
gam_order()
}
\value{
A list corresponding to assigned gamete order.
}
\description{
For internal use.
}
\examples{
gam_order()

}

