% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_functions.R
\name{genos2sparseMatrix}
\alias{genos2sparseMatrix}
\title{Convert genotypes to sparse Matrix format.}
\usage{
genos2sparseMatrix(genotypes)
}
\arguments{
\item{genotypes}{The genotypes returned by read.vcfR}
}
\value{
A sparseMatrix
}
\description{
Convert genotypes to sparse Matrix format.
}
