% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_Functions.R
\name{get_FGenos}
\alias{get_FGenos}
\title{Set Founder Genotypes and Return Reduced DF of Genotypes}
\usage{
get_FGenos(founder_ids, RV_founder, FamID, founder_genotypes, FamRV)
}
\value{
list of (1) family genotypes and (2) a reduced pool of genos
}
\description{
Set Founder Genotypes and Return Reduced DF of Genotypes
}
