% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_Functions.R
\name{get_parOffInfo}
\alias{get_parOffInfo}
\title{Get parent and offspring information from a pedigree}
\usage{
get_parOffInfo(ped_file)
}
\arguments{
\item{ped_file}{Data.frame. The pedigree file, must have same format as pedigree simulated with \code{sim_RVped}}

\item{offspring_ID}{Numeric. The ID of the offspring.}

\item{parent_ID}{Numeric. The ID of the parent.}
}
\value{
A list containing the parent's paternal and maternal alleles at the disease locus, and the RV status of the offspring
}
\description{
For internal use.
}

