% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_functions.R
\name{get_SMindex_by_person}
\alias{get_SMindex_by_person}
\title{Returns the row and columns locations of mutations for a person}
\usage{
get_SMindex_by_person(person_index, person_genos)
}
\arguments{
\item{person_index}{The column location of the person}

\item{person_genos}{The genotype vector for the person}
}
\value{
A list containing two vectors, the first contains the row locations of mutations, the second contains the column locations of mutations.
}
\description{
Returns the row and columns locations of mutations for a person
}
\keyword{internal}
