% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg_chrom.R
\docType{data}
\name{hg_chrom}
\alias{hg_chrom}
\title{Human Chromosome Dataset}
\format{A data set with 2 rows and 4 variables:
\describe{
  \item{chrom}{Numeric. The chromosome number.}
  \item{start_pos}{Numeric. The starting position, in bp, over which to simulate data.}
  \item{end_pos}{Numeric. The ending position, in bp, over which to simulate data.}
 \item{center}{Numeric. The position of the centromere, in bp.}
}}
\usage{
hg_chrom
}
\description{
Example chromosome data frame, contains data for the 22 human autosomes.
}
\keyword{datasets}
