% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerMap_Methods.R
\name{identify_pathwayRVs}
\alias{identify_pathwayRVs}
\title{Identify variants located in a defined pathway}
\usage{
identify_pathwayRVs(path_by_chrom, marker_map_by_chrom)
}
\arguments{
\item{path_by_chrom}{The pathway data for the chromosome under consideration}

\item{marker_map_by_chrom}{The marker_map for the chromosome under consideration}
}
\value{
marker_map_by_chrom with possibleRV identified
}
\description{
Identify variants located in a defined pathway
}
\keyword{internal}
