% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyPathwaySNVs.R
\name{identify_pathwaySNVs}
\alias{identify_pathwaySNVs}
\title{Identify SNVs located in a specified pathway}
\usage{
identify_pathwaySNVs(markerDF, pathwayDF)
}
\arguments{
\item{markerDF}{A data frame containing SNV data.   See details.}

\item{pathwayDF}{A data frame containing pathway data.  See details.}
}
\value{
A dataframe with the same format as \code{markerDF}, but with a variable \code{pathwaySNV} marked FALSE for all variants located outside exons in \code{pathwayDF}, or with derived allele frequency greater than \code{carrier_prob}.
}
\description{
Identify SNVs located in a specified pathway
}
\keyword{internal}
