% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerMap_Methods.R
\name{identify_possibleRVs}
\alias{identify_possibleRVs}
\title{Identify possible variants based on defined pathway}
\usage{
identify_possibleRVs(path_by_chrom, marker_map_by_chrom)
}
\arguments{
\item{path_by_chrom}{The pathway data for the chromosome under consideration}

\item{marker_map_by_chrom}{The marker_map for the chromosome under consideration}
}
\value{
marker_map_by_chrom with possibleRV identified
}
\description{
Identify possible variants based on defined pathway
}
\keyword{internal}
