% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerMap_Methods.R
\name{is.markerMap}
\alias{is.markerMap}
\title{Check to see if object is of class markerMap}
\usage{
is.markerMap(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
Logical. Indicates if \code{x} is of class \code{markerMap}.
}
\description{
Check to see if object is of class markerMap
}
\keyword{internal}
