% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_Functions.R
\name{is_int}
\alias{is_int}
\title{Determine if input is an integer}
\usage{
is_int(x)
}
\arguments{
\item{x}{Numeric.}
}
\value{
Boolean
}
\description{
Determine if input is an integer
}
\keyword{internal}
