% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_Functions.R
\name{is_odd}
\alias{is_odd}
\title{Determine if input is an odd number}
\usage{
is_odd(x)
}
\arguments{
\item{x}{Numeric.}
}
\value{
Boolean
}
\description{
Determine if input is an odd number
}
\keyword{internal}
