% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_map.R
\docType{data}
\name{mark_map}
\alias{mark_map}
\title{Marker Map Dataset}
\format{A data set with 14 rows and 4 variables:
\describe{
  \item{chrom}{Numeric. The chromosome on which the marker resides.}
  \item{position}{The location of the marker, in basepairs.}
  \item{pathwayID}{A pathwayID for the variant.}
 \item{possibleRV}{Logical. Indicates whether the variant may be chosen as a causal familial variant.}
}}
\usage{
mark_map
}
\description{
Example marker map data frame, used to illustrate usage of the \code{\link{markerMap}} function.
}
\keyword{datasets}
