% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerMap_Methods.R
\name{markerMap}
\alias{markerMap}
\title{Create an object of class markerMap.}
\usage{
markerMap(markerDF, pathwayDF = NULL)
}
\arguments{
\item{markerDF}{A data frame containing pertinent information on markers.   See details.}
}
\value{
An object of class markerMap.
}
\description{
Create a markerMap object, required input for \code{\link{sim_RVstudy}} and \code{\link{sim_RVseq}} functions.
}
\details{
The named columns in the data frame, \code{markerDF}, are described as follows:
\describe{
 \item{marker}{(OPTIONAL) An identifying name/number for the variant. (If missing this field will be automatically generated from the columns named "chrom" and "position" as "chrom_position"). NOTE: These names must match \emph{exactly} the column names (i.e. the marker names) in the haplotype_dist dataframe.}
 \item{chrom}{The number of the chromosome on which the marker resides.}
 \item{position}{The location of the variant, in basepairs.}
 \item{possibleRV}{(OPTIONAL) Logical variable, indicates whether the variant may be chosen as a causal familial variant.  If this field is missing and \code{pathwayDF} is not provided it is assumed that all variants listed are candidates for selection.}
 \item{probCausal}{(OPTIONAL) The probability that the variant is selected as a familial RV.  If this field is missing, it will be assumed that all variants for which possibleRV is TRUE are equally likely.}
}

TO FIX: CURRENTLY NOT ACCOUNTING FOR CARRIER PROB WHEN CHOOSING VARIANTS
}
\examples{
data(mark_map)

head(mark_map)
class(mark_map)

markObj <- markerMap(mark_map)
head(markObj)
class(markObj)

}
