% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_arlequinFunctions.R
\name{read.arp}
\alias{read.arp}
\title{Read arlequin project file}
\usage{
read.arp(file)
}
\arguments{
\item{file}{character.  The name of the arlequin project file, note that the file name must include the .arp file extension.}
}
\value{
A list containing:

\item{\code{SNPs} }{A dataframe of SNPs, with columns for each loci in Postions and rows for individual haplotypes.}

\item{\code{Positions} }{A numeric list of the postions, in bp, of the SNPs contained in SNPs}
}
\description{
Read arlequin project file with DataType SNP only, and extract the SNPs as a dataframe well as their postions.
}
\examples{
\dontrun{
setwd("C:/Users/cnieuwoudt/Documents/fsc26_win64/1PopDNAtest")
exDat = read.arp("1PopDNAtest_1_1.arp")
head(exDat[[1]])
head(exDat[[2]])
}

}
