% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlimFunctions.R
\name{read_slim}
\alias{read_slim}
\title{Read SLiM 2.0 Output}
\usage{
read_slim(file_path, keep_maf = 0.01, recomb_map = NULL)
}
\arguments{
\item{file_path}{character.  The file path of the SLiM 2.0 output file.}

\item{keep_maf}{numeric. The largest allele frequency for retained SNPs.  All variants with allele frequency greater than \code{keep_maf} will be removed.  Please note, removing common variants is recommended for large datasets due to the limitations of data allocation in \code{R}.}

\item{recomb_map}{data frame. The recombination map provided to SLiM 2.0, typically created by}
}
\value{
A list containing:

\item{\code{Mutations} }{A dataframe containing SNP information, see details..}

\item{\code{Genomes} }{A sparse matrix of haplotypes, see details.}
}
\description{
Extract SNP data from SLiM output.
}
\details{
The \code{read_slim} function is used to extract SNP (single nucleotide polymorphism) data from SLiM model output text file.

The first item returned by \code{read_slim} is a data frame named \code{Mutations}, which contains information regarding SNP IDs, postions, and allele frequency.  The variable \code{colID} references the the column position (in the sparse genotype matrix) of the SNP, \code{position} is the genomic position of the SNP (in bp), and \code{afreq} is the allele frequency of the SNP.

The second item returned is a sparse matrix named \code{Genomes}.  This matrix is the collection of haplotypes for each individual in the population.  Each row represents one of the haplotypes for a single individual in the population.  For example, the first individual's maternally and paternally inherited haplotypes are stored in rows one and two, respectively.  In general, the \eqn{i^{th}} individual's maternally inherited haplotype is stored in row \eqn{2i-1} and the paternally inherited haplotype is stored in row \eqn{2i}. (ALTERNATIVE DESC: The third and fourth rows contain haplotypes for the second individual, fifth and sixth rows contain haplotypes for the second individual, and so on. )

NOTE TO SELF:
\itemize{
 \item For file extension internal check use: importFrom tools file_ext
 \item If possible, add test: two rows for each person in genotypes matrix
}
}
\examples{
sout = read_slim(file_path = "C:/Data/Slim/SlimFull_output.txt", keep_maf = 0.01)
sout = read_slim(file_path = "C:/Data/Slim/SLiMtest_output.txt", keep_maf = 0.01)
}
