% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_Functions.R
\name{reduce_to_events}
\alias{reduce_to_events}
\title{Remove chiasmata locations that a gamete is not involved in.}
\usage{
reduce_to_events(gamete_haplo, chias_locations)
}
\arguments{
\item{gamete_haplo}{Numeric vector. The inherited haplotype.}

\item{chias_locations}{Numeric vector.  Chiasmata locations.}
}
\value{
The locations of crossovers
}
\description{
Remove chiasmata locations that a gamete is not involved in.
}

