% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlimMapFunctions.R
\name{reMap_mutations}
\alias{reMap_mutations}
\title{Re-map slim mutations}
\usage{
reMap_mutations(mutationDF, genoMat, recombMap)
}
\arguments{
\item{mutationDF}{The Mutation data frame returned by reMap_mutations}

\item{genoMat}{The sparse genotypes matrix returned by reMap_mutations}

\item{recombMap}{The recombination map provided to slim}
}
\value{
A re-mapped mutation data frame
}
\description{
Re-map slim mutations
}
