% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_StudySeqFunctions.R
\name{remove_allWild}
\alias{remove_allWild}
\title{Remove markers at which all individuals carry wild type allele}
\usage{
remove_allWild(f_haps, marker_map)
}
\arguments{
\item{f_haps}{The founder haplotypes data. This is a list of lists (by family). By family, this contains the haplotypes for each founder (first item), and a list of ID numbers (second item) which is used to map the haplotype to the person to whom it belongs.}

\item{marker_map}{data.frame. Catalogs the SNV data contained in the familial haplotypes.}
}
\value{
A list (by family) of haplotype matrices and ID vectors and the reduce marker data set.
}
\description{
Remove markers at which all individuals carry wild type allele
}
\examples{
#probably no examples

}
