% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_StudySeqFunctions.R
\name{resample}
\alias{resample}
\title{Dummy sample function}
\usage{
resample(x, ...)
}
\arguments{
\item{x}{A vector from which to sample}

\item{...}{All other arguments in sample}
}
\value{
The randomly sampled items
}
\description{
Dummy sample function
}
\keyword{internal}
