% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gameteFunctions.R
\name{sim_chiasmataPositions}
\alias{sim_chiasmataPositions}
\title{Simulate chiasmata positions}
\usage{
sim_chiasmataPositions(chrom_map, mean_dist = 0.5, gamma_params = c(2.63,
  2.63/0.5))
}
\arguments{
\item{chrom_map}{Data.frame with 1 row and 2 columns. The two columns represent the start and stop positions (in cM) over which to simulate recombination.}

\item{gamma_params}{Numeric list of length 2. The respective shape and scale parameters gamma distribution used to simulate distance between chiasmata, default from Vorrips.}
}
\value{
A list containing chiasmata postions and the number of chiasmata that occur before the centromere.
}
\description{
Simulate chiasmata positions among a group of four chromatids.
}
\examples{

sim_chiasmataPositions(chrom_map = data.frame(start = 47, stop = 198, center = 100))

set.seed(1)
system.time(for(i in 1:10000){
sim_chiasmataPositions(chrom_map = data.frame(start = 0, stop = 200, center = 50))
})


}

