% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_functions.R
\name{sim_chiasmataPositionsNI}
\alias{sim_chiasmataPositionsNI}
\title{Simulate chiasmata positions}
\usage{
sim_chiasmataPositionsNI(chrom_map, mean_dist = 0.5)
}
\arguments{
\item{chrom_map}{Data.frame with 1 row and 2 columns. The two columns represent the start and stop positions (in cM) over which to simulate recombination.}

\item{mean_dist}{Numeric. The mean of the expoential distribution.}
}
\value{
A list containing chiasmata postions.
}
\description{
Simulate chiasmata positions
}
\details{
Issues: from what exponential distribution should the original start pos come from? Currently using exponential with mean 0.5

Also, need to verify that parmaters chosen for gamma distribution result in an average distance of 1 crossover per M.
}
\examples{

sim_chiasmataPositionsNI(chrom_map = data.frame(start = 47, stop = 198, center = 100))

set.seed(1)
system.time(for(i in 1:10000){
sim_chiasmataPositionsNI(chrom_map = data.frame(start = 0, stop = 200, center = 50))
})


}

