% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_StudySeqFunctions.R
\name{sim_FGenos}
\alias{sim_FGenos}
\title{Draw Founder Genotypes from Haplotype Distribution Given Familial Risk Variant}
\usage{
sim_FGenos(founder_ids, RV_founder, haplotype_dist, RV_col_loc)
}
\arguments{
\item{founder_ids}{Numeric list. The ID numbers of all non-seed founders.}

\item{RV_founder}{Numeric. The ID number of the seed founder.}

\item{haplotype_dist}{sparseMatrix.  The sparseMatrix of genomes returned by \code{read_slimOut}.}

\item{RV_col_loc}{Nueric. The column location of the familial RV in haplotype_dist.}
}
\value{
list of familial founder genotypes
}
\description{
\strong{For internal use.}
}
