% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gameteFunctions.R
\name{sim_gameteFormation}
\alias{sim_gameteFormation}
\title{Simulate gamete formation}
\usage{
sim_gameteFormation(chrom_map, allele_IDs)
}
\arguments{
\item{chrom_map}{Data.frame.  A data.frame consisting of three columns: column 1 contains the chromosome numbers, column 2 start postion of chromosome (in cM), column 3 end position of chromosome (in cM).}

\item{allele_IDs}{Numeric list of length 2. The identification numbers for the respective paternal and maternal alleles of the individual for whom we wish to simulate recombination.}
}
\value{
chrom_haps and gamete_group
}
\description{
Simulate gamete formation
}
\examples{
my_chrom_map = data.frame(chrom = c(1, 2, 3),
                          start_pos = c(0, 0, 0),
                          end_pos = c(250, 120, 75),
                          center = c(50, 24, 15))

sim_gameteFormation(my_chrom_map, allele_IDs = c(0, 1))

system.time(for (i in 1:10000) {
  sim_gameteFormation(my_chrom_map, c(0, 1))
})
}

