% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gameteFunctions.R
\name{sim_haploidFormation}
\alias{sim_haploidFormation}
\title{Simulate recombination along a bundle of four chromatids.}
\usage{
sim_haploidFormation(num_chiasmata, before_center, allele_IDs)
}
\arguments{
\item{num_chiasmata}{Numeric. The number of chiasmata to simulate among the chromatid bundle.}

\item{allele_IDs}{Numeric list of length 2. The identification numbers for the respective paternal and maternal alleles of the individual for whom we wish to simulate recombination.}
}
\value{
haploid_mat. A matrix with rows representing recombined haplotypes.
}
\description{
Simulate recombination along a bundle of four chromatids.
}
\details{
Given the possible chiasmata positions returned from \code{sim_chiasmataPostions}, we want to randomly select two non-sister chromatids to participate in each recombination event.  We assume no chromatid interference so that the non-sister chromatids participating in a crossover event are independent of those chosen in previous crossover events.
}
\examples{
my_chrom_map <- data.frame(start = 0, stop = 200, center = 40)
sim_chias_pos <- sim_chiasmataPositions(my_chrom_map)
chias_count_BC(sim_chias_pos, 40)
sim_haploidFormation(num_chiasmata = length(sim_chias_pos),
                     before_center = chias_count_BC(sim_chias_pos, my_chrom_map[1,3]),
                     allele_IDs = c(2, 3))

system.time(for (i in 1:10000) {
sim_chias_pos <- sim_chiasmataPositions(my_chrom_map)
sim_haploidFormation(num_chiasmata = length(sim_chias_pos),
                     before_center = chias_count_BC(sim_chias_pos, my_chrom_map[1,3]),
                     allele_IDs = c(2, 3))
})

}
\seealso{
\code{\link{sim_chiasmataPostions}}
}

