% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_StudySeqFunctions.R
\name{sim_RVstudy}
\alias{sim_RVstudy}
\title{Simulate sequence data for a study}
\usage{
sim_RVstudy(ped_files, marker_map, chrom_map, haplotype_dist,
  affected_only = TRUE, convert_to_cM = TRUE, burn_in = 1000,
  gamma_params = c(2.63, 2.63/0.5))
}
\arguments{
\item{ped_files}{Data frame. Must match format of pedigree simulated by sim_RVped}

\item{marker_map}{Data.frame. Must contain three columns with: column 1: marker names, must be listed in the same order as in the founder genotype file, column 2: the chromosomal position of the marker, column 3: the position of the marker in cM.}

\item{chrom_map}{Data.frame.  A data.frame consisting of three columns: column 1 contains the chromosome numbers, column 2 start postion of chromosome (in cM), column 3 end position of chromosome (in cM).}

\item{burn_in}{Numeric. The burn in distance in cM.}

\item{gamma_params}{Numeric list of length 2. The respective shape and rate parameters gamma distribution used to simulate distance between chiasmata.  By default, \code{gamma_params = c(2.63, 2.63/0.5)}, as discussed in Vorrips (2012).}

\item{haplotype_dist.}{List of dataframes.  Each list entry should contain a dataframe containing the haplotype distribution for that chromosome (in same order as marker_map).}

\item{founder_genotypes}{list of dataframes. Should contain, in the order listed in chrom_map, 1 dataframe for every chromosome contained in chrom_map.  Each data frame should contain two rows for each founder: the first to correspond to the paternally inherited gamete and the second to correspond to the maternally inherited gamete.}

\item{RV_markers}{character. A list of possible RV markers. If missing all markers in \code{marker_map} are assumed to be possible RV markers.}
}
\value{
study_sequences
}
\description{
Simulate sequence data for a study
}
\examples{
library(SimRVPedigree)
data(EgPeds)

EgPeds5 <- EgPeds4 <- EgPeds3 <- EgPeds2 <- EgPeds
EgPeds2$FamID <- EgPeds2$FamID + 5
EgPeds3$FamID <- EgPeds3$FamID + 10
EgPeds4$FamID <- EgPeds4$FamID + 15
EgPeds5$FamID <- EgPeds5$FamID + 20
ex_study_peds <- rbind(EgPeds, EgPeds2, EgPeds3, EgPeds4, EgPeds5)
ex_study_peds <- EgPeds

nrow(ex_study_peds[which(is.na(ex_study_peds$dadID)), ])
length(unique(ex_study_peds$FamID))

my_chrom_map = data.frame(chrom     = c(1, 2),
                          start_pos = c(0, 0),
                          end_pos   = c(247199719, 242751149),
                          center = c(98879888, 97100460))
my_chrom_map

data(mark_map)
head(mark_map)

mm_obj <- markerMap(mark_map)
head(mm_obj)

set.seed(1)
founder_seq <- as.data.frame(matrix(sample(2*nrow(mm_obj)*1000,
                                    x = c(0, 1),
                                    replace = T,
                                    prob = c(0.95, 0.05)),
                             nrow = 2*1000))
colnames(founder_seq) = as.character(mm_obj$marker)
hdist = estimate_haploDist(founder_seq, mm_obj)

hdist2 = estimate_haploDist(founder_seq[which(founder_seq[, 2] == 1), ], mm_obj)
hdist2[[1]]

condition_haploDist(hdist[[1]], "1_5012369", 1)

set.seed(6)
ped_seq <- sim_RVstudy(ped_files = ex_study_peds,
                       haplotype_dist = estimate_haploDist(founder_seq, mm_obj),
                       marker_map = mm_obj,
                       chrom_map = my_chrom_map)
ped_seq

set.seed(6)
system.time(sim_RVstudy(ped_files = ex_study_peds,
                        haplotype_dist = estimate_haploDist(founder_seq, mm_obj),
                        marker_map = mm_obj,
                        chrom_map = my_chrom_map))

}
