% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_dat.R
\docType{data}
\name{SNP_dat}
\alias{SNP_dat}
\title{SNP Dataset}
\format{A data set with 10000 rows and 366 variables:
\describe{
 \item{17_11}{Numeric. The snps at locus 11.}
 \item{17_288}{Numeric. The snps at locus 288.}
 \item{17_635}{Numeric. The snps at locus 635.}
 ...
}}
\usage{
SNP_dat
}
\description{
Example dataframe of SNPs, corresponds to the data in mark_map.
}
\keyword{datasets}
