% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_seq_methods.R
\name{summary.famStudy}
\alias{summary.famStudy}
\title{Summary function for objects of class famStudy}
\usage{
\method{summary}{famStudy}(object, ...)
}
\arguments{
\item{object}{An object of class \code{famStudy}, returned by the \code{sim_RVstudy} function.}

\item{...}{additional arguments passed to other methods.}
}
\value{
Not sure yet..
}
\description{
Summary function for objects of class famStudy
}
\examples{
library(SimRVSequences)

#load pedigree, haplotype, and mutation data
data(study_peds)
data(EXmuts)
data(EXhaps)

#create variable is_CRV in EXmuts to identify causal
#rare variants, from which to sample familial variants.
EXmuts$is_CRV = FALSE
EXmuts$is_CRV[c(2, 3, 12, 24)] = TRUE

#supply required inputs to the sim_RVstudy function
seqDat = sim_RVstudy(ped_files = study_peds,
                     SNV_map = EXmuts,
                     haplos = EXhaps)

#to count the number of SNVs shared by the disease-affected
#relatives in each pedigree, supply the output returned by
#sim_RVstudy to the summary function
summary(seqDat)


}
