% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_Functions.R
\name{affected_onlyPed}
\alias{affected_onlyPed}
\title{Remove unaffected relatives}
\usage{
affected_onlyPed(ped_file)
}
\arguments{
\item{ped_file}{data.frame. A pedigree.}
}
\value{
\code{retA_ped} A pedigree containing only affected members, obligate carriers, and founders.
}
\description{
Remove unaffected relatives
}
\examples{
library(SimRVPedigree)
#Read in example pedigrees and create ped object
data(EgPeds)
ex_peds <- new.ped(EgPeds)

#plot full pedigree
plot(ex_peds[which(ex_peds$FamID == 1), ])

#reduce to affected only pedigree
Apeds = affected_onlyPed(ex_peds[which(ex_peds$FamID == 1), ])
plot(Apeds)

}
