% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_peds}
\alias{check_peds}
\title{Checks ped_files for expected info and format.}
\usage{
check_peds(ped_files)
}
\arguments{
\item{ped_files}{Data frame. A data frame of pedigrees for which to simulate sequence data, see details.}
}
\description{
\strong{For internal use.}
}
\keyword{internal}
