% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_SNV_map}
\alias{check_SNV_map}
\title{Check SNV_map for possible issues}
\usage{
check_SNV_map(SNV_map)
}
\arguments{
\item{SNV_map}{Data frame. A data frame that catalogs the SNVs in \code{haplos}.  If the \code{\link{read_slim}} function was used to import SLiM data to \code{R}, the data frame \code{Mutations} is of the proper format for \code{SNV_map}.  However, users must add the variable \code{is_CRV} to this data frame, see details.}
}
\description{
\strong{For internal use.}
}
\keyword{internal}
