% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_SNV_map}
\alias{check_SNV_map}
\title{Check SNV_map for possible issues}
\usage{
check_SNV_map(SNV_map)
}
\arguments{
\item{SNV_map}{This argument has been deprecated. Users now supply objects of class \code{SNVdata} to argument \code{SNV_data}.}
}
\description{
\strong{For internal use.}
}
\keyword{internal}
