% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_exons.R
\name{combine_exons}
\alias{combine_exons}
\title{Combine overlapping exons into a single observation}
\usage{
combine_exons(exon_data, zero_start = FALSE)
}
\arguments{
\item{exon_data}{The exon data to combine.  This data must unclude named variables: 'chrom', a chromosome identifer; 'exonStart', the first position of the exon in base pairs; and 'exonEnd', the last position of the exon in base pairs.}

\item{zero_start.}{Logical.  \code{zero_start = TRUE} if the first position is 0-based, and \code{zero_start = FALSE} if the first position is 1-based.}
}
\value{
A matrix of exon unions, i.e. non-overlapping observations, by chromosome.
}
\description{
Combine overlapping exons into a single observation
}
\examples{
# create an example data frame that contains the
# the variables: chrom, exonStart, and exonEnd
exDat <- data.frame(chrom     = c(1, 1, 1, 2, 2, 2),
                    exonStart = c(1, 2, 5, 1, 3, 3),
                    exonEnd   = c(3, 4, 7, 4, 5, 6))

exDat

# supply exDat to combine_exons to combine overlapping segments
# into a single observation.
combine_exons(t1)

}
