% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_Functions.R
\name{condition_haploDist}
\alias{condition_haploDist}
\title{Condition Haplotype distribution}
\usage{
condition_haploDist(Chaplo_dist, RV_marker, RV_status)
}
\arguments{
\item{RV_status}{}
}
\description{
Condition Haplotype distribution
}

