% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_seq_methods.R
\name{count_affectedRV}
\alias{count_affectedRV}
\title{Summary function for object of class fam_study}
\usage{
count_affectedRV(fam_study)
}
\arguments{
\item{fam_study}{The object returned by sim_RVstudy}
}
\value{
Not sure yet..
}
\description{
Summary function for object of class fam_study
}
\examples{
#create example
}
