% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_map_functions.R
\name{create_chrom_map}
\alias{create_chrom_map}
\title{Create chromosome map from marker map}
\usage{
create_chrom_map(SNV_map)
}
\arguments{
\item{SNV_map}{Data frame. A data frame that catalogs the SNVs in \code{haplos}.  If the \code{\link{read_slim}} function was used to import SLiM data to \code{R}, the data frame \code{Mutations} is of the proper format for \code{SNV_map}.  However, users must add the variable \code{is_CRV} to this data frame, see details.}
}
\value{
a dataframe catalouging the start and stop positions, in base pairs, for each chromosome.  We use this information to determine what regions to simulate recombination over.
}
\description{
INTENDED FOR INTERNAL USE ONLY
}
\keyword{internal}
