% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlimFunctions.R
\name{create_slimMap}
\alias{create_slimMap}
\title{Create recombination map}
\usage{
create_slimMap(exon_df)
}
\arguments{
\item{exon_df}{}
}
\value{
rc_map A recombination map that may be used in conjunction with SLiM 2.0 (cite Messer), provided that the end position is shifted forward by one position.  See example.
}
\description{
Create a recombination map for use with SLiM 2.0 (Messer)
}
\details{
Why would we do this? what is it used for?

#' NOTE TO SELF:
\itemize{
 \item Add muation_rate and recombination_rate parameters
 \item change exon_df to more general description
}
}
\examples{
data(hg_exons)
smap <- create_slimMap(hg_exons)
head(smap)

NOTE: The first posistion in an eidos array begins at zero,
hence, to use with Slim, users must make the following change
smap$endPos <- smap$endPos - 1
head(smap)

}
