% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_Functions.R
\name{estimate_haploDist}
\alias{estimate_haploDist}
\title{Estimate haplotype distribution}
\usage{
estimate_haploDist(pop_haplos, marker_map)
}
\arguments{
\item{pop_haplos}{Data.frame.  A data frame containing the population haplotypes}

\item{marker_map}{Data.frame. Must contain three columns with: column 1: marker names, must be listed in the same order as in the founder genotype file, column 2: the chromosomal position of the marker, column 3: the position of the marker in cM.}
}
\value{
haplo_dist The haplotype distribution
}
\description{
Estimate haplotype distribution
}
\examples{
mark_map <- data.frame(chrom = c(1, 1, 1, 1, 1, 1, 1, 1,
                                      2, 2, 2, 2, 2, 2),
                       position = c(5012368, 5012369, 5012370,
                                    78541008, 78541009, 78541010,
                                    247199219, 247199220,
                                    11330, 11332,
                                    234577, 234578, 234579,
                                    18799180),
                       pathwayID = c(1, 1, 1, 2, 2, 2, 3, 3,
                                     2, 2, 3, 3, 3, 1),
                       possibleRV = c(0, 1, 1, 0, 0, 0, 1, 1,
                                      0, 0, 0, 1, 1, 1))
mark_map$marker <- paste0(mark_map$chrom, sep = "_", mark_map$position)
mark_map <- mark_map[, c(5, 1:4)]
mark_map

set.seed(1)
founder_seq <- as.data.frame(matrix(sample(2*nrow(mark_map)*1000,
                                    x = c(0, 1),
                                    replace = T,
                                    prob = c(0.95, 0.05)),
                             nrow = 2*1000))
colnames(founder_seq) = as.character(mark_map$marker)

hdist = estimate_haploDist(founder_seq, mark_map)
hdist[[1]]
hdist[[2]]


}

