% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXhaps.R
\docType{data}
\name{EXhaps}
\alias{EXhaps}
\title{Example Haplotypes Dataset}
\format{A sparseMatrix of class dgCMatrix with 20000 rows and 500 variables.  Each row represents a haplotype and each column represents an SNV locus.}
\usage{
EXhaps
}
\description{
This data set contains 20,000 haplotypes (i.e. rows) spanning 500 single-nucleotide variants.
}
\details{
This dataset is intended to accompany the \code{EXmuts} dataset and is used for demonstration purposes. For additional information regarding the SNVs contained in \code{EXhaps} please refer to the documentation for the \code{\link{EXmuts}} data set.
}
\seealso{
\code{\link{EXmuts}}, \code{\link{read_slim}}
}
\keyword{datasets}
