% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXmut.R
\docType{data}
\name{EXmut}
\alias{EXmut}
\title{Example Mutations Dataset}
\format{A data set with 173 rows and 4 variables:
\describe{
  \item{colID}{Numeric. The corresponding column number of the SVN in the EXgen dataset.}
  \item{chrom}{Numeric. The choromosome number.}
  \item{position}{Numeric. The location of the SNV, in base pairs.}
  \item{afreq}{Numeric. The derived allele frequency of the SNV.}
  \item{marker}{Character. The names of the genes contained in the combined exon.}
}}
\usage{
EXmut
}
\description{
This data set contains mutations along segments of length 20,000 base pairs from human chromosomes 1-5. This dataset is intended to accompany the EXgen dataset and is used for demonstration purposes.
}
\seealso{
EXgen
}
\keyword{datasets}
