% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXmuts.R
\docType{data}
\name{EXmuts}
\alias{EXmuts}
\title{Example Mutations Dataset}
\format{A data set with 450 rows and 6 variables:
\describe{
  \item{colID}{Numeric. The corresponding column number of the SVN in the EXgen dataset.}
  \item{chrom}{Numeric. The chromosome number.}
  \item{position}{Numeric. The location of the SNV, in base pairs.}
  \item{afreq}{Numeric. The derived allele frequency of the SNV.}
  \item{marker}{Character. The names of the genes contained in the combined exon.}
  \item{pathwaySNV}{Logical. Indicates if the SNV is located within the pathway.}
}}
\usage{
EXmuts
}
\description{
This data set contains mutations along segments of length 20,000 base pairs from human chromosomes 1-5. This dataset is intended to accompany the EXgen dataset and is used for demonstration purposes.
}
\seealso{
EXgen
}
\keyword{datasets}
