% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlimFunctions.R
\name{extract_tempIDs}
\alias{extract_tempIDs}
\title{Determine i and j positions of mutations for sparse matrix}
\usage{
extract_tempIDs(mutString, rarePos)
}
\arguments{
\item{mutString}{character. String containing mutations}

\item{rarePos}{numeric vector. position of variations}
}
\value{
data.frame with x and y positions of mutations
}
\description{
Determine i and j positions of mutations for sparse matrix
}
\keyword{internal}
