% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_functions.R
\name{genos2sparseMatrix}
\alias{genos2sparseMatrix}
\title{Convert genotypes to haplotypes.}
\usage{
genos2sparseMatrix(genotypes)
}
\arguments{
\item{genotypes}{The genotypes returned by read.vcfR, see details.}
}
\value{
A sparseMatrix
}
\description{
This function may be used to convert phased genotype data for diplod organisms into a sparse matrix in which each row represents the phased haplotype data for an individual.
}
