% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVdata_methods.R
\name{import_SNVdata}
\alias{import_SNVdata}
\title{Import 1000 genomes exon data formatted for sim_RVstudy}
\usage{
import_SNVdata(chrom)
}
\arguments{
\item{chrom}{The chromosome number.  Can be a numeric list of chromosome numbers or a character argument set to \code{"ALL"} if data from all chromosomes should be imported.}
}
\value{
An object of class \code{SNVdata} or a list of objects of class \code{SNVdata}, i.e. one for each chomosome that was imported.
}
\description{
Import 1000 genomes exon data formatted for sim_RVstudy
}
\examples{
exdata = import_SNVdata(0)

head(exdata$Mutations)
exdata$Haplotypes[1:20, 1:10]
}
