% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_Functions.R
\name{reduce_to_events}
\alias{reduce_to_events}
\title{Reduce chiasmata vector to crossovers that transmitted gamete participates in based on the allele vector.}
\usage{
reduce_to_events(gamete_haplo, chias_locations)
}
\arguments{
\item{gamete_haplo}{Numeric vector. The inherited haplotype.}

\item{chias_locations}{Numeric vector.  Chiasmata locations.}
}
\value{
The locations of crossovers
}
\description{
For internal use.
}
\examples{
\dontrun{
haplo_vec <- sample(x = c(2, 3), size = 10, replace = TRUE)
chias_vec <- cumsum(rgamma(9, shape = 2.63, rate = 2*2.63))
reduce_to_events(gamete_haplo = haplo_vec,
                 chias_locations = chias_vec)
}
}
\keyword{internal}
