% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlimFunctions.R
\name{reMap_mutations}
\alias{reMap_mutations}
\title{Re-map slim mutations}
\usage{
reMap_mutations(mutationDF, recomb_map)
}
\arguments{
\item{mutationDF}{The Mutation data frame returned by reMap_mutations}

\item{recomb_map}{The recombination map provided to slim}
}
\value{
A re-mapped mutation data frame
}
\description{
Intended for internal use
}
\keyword{internal}
