% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_StudySeqFunctions.R
\name{remove_allWild}
\alias{remove_allWild}
\title{Remove unmutated markers from data}
\usage{
remove_allWild(f_haps, SNV_map)
}
\arguments{
\item{f_haps}{The founder haplotypes data. This is a list of family lists. By family, this contains the haplotypes for each founder (first item), and a list of ID numbers (second item) which is used to map the haplotype to the person to whom it belongs.}

\item{SNV_map}{data.frame. Catalogs the SNV data contained in the familial haplotypes.}
}
\value{
A list (by family) of haplotype matrices and ID vectors and the reduce marker data set.
}
\description{
Remove any markers for which all founders, in the study, are homozygous for the wild-type allele.  Since we do not model de novo mutations, it is not possible for non-founders to develop mutations at these loci.
}
\keyword{internal}
