% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chiasmata_SimFunctions.R
\name{sim_chiasmataPositions}
\alias{sim_chiasmataPositions}
\title{Simulate chiasmata positions}
\usage{
sim_chiasmataPositions(chrom_length, gamma_params = c(2.63, 2.63/2))
}
\arguments{
\item{chrom_length}{Numeric. The chromosome length, in M, over which to simulate chiasmata.}

\item{gamma_params}{Numeric list of lenght 2. The respective shape and scale parameters gamma distribution used to simulate distance between chiasmata, default from Vorrips.}
}
\value{
A vector of chiasmata postions.
}
\description{
Simulate chiasmata positions
}
\details{
Issues: from what exponential distribution should the original start pos come from? Currently using exponential with mean 0.5
}
\examples{

sim_chiasmataPositions(chrom_length = 25)

set.seed(1)
system.time(for(i in 1:10000){
sim_chiasmataPositions(chrom_length = 500)
})

}

