% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_Functions.R
\name{sim_FGenos}
\alias{sim_FGenos}
\title{Draw Founder Genotypes from Haplotype Distribution Given Familial Risk Variant}
\usage{
sim_FGenos(founder_ids, RV_founder, FamID, haplotype_dist, FamRV, marker_map)
}
\value{
list of familial founder genotypes
}
\description{
Draw Founder Genotypes from Haplotype Distribution Given Familial Risk Variant
}
