% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_StudySeqFunctions.R
\name{sim_FGenos}
\alias{sim_FGenos}
\title{Draw Founder Genotypes from Haplotype Distribution Given Familial Risk Variant}
\usage{
sim_FGenos(founder_ids, RV_founder, RV_founder_pat, haplos, RV_col_loc,
  RV_pool_loc)
}
\arguments{
\item{founder_ids}{Numeric list. The ID numbers of all non-seed founders.}

\item{RV_founder}{Numeric. The ID number of the seed founder.}

\item{RV_founder_pat}{Numeric. RV_founder_pat == 1 if RV founder inherited the RV from dad, and 0 if inherited RV from mom.}

\item{haplos}{sparseMatrix.  The sparseMatrix of genomes returned by \code{read_slimOut}.}

\item{RV_col_loc}{Numeric. The column location of the familial RV in haplos.}

\item{RV_pool_loc}{The column locations of each SNV in the pool of candidate SNVs.}
}
\value{
list of familial founder genotypes
}
\description{
\strong{For internal use.}
}
\keyword{internal}
