% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_StudySeqFunctions.R
\name{sim_FGenos}
\alias{sim_FGenos}
\title{Draw Founder Genotypes from Haplotype Distribution Given Familial Risk Variant}
\usage{
sim_FGenos(founder_ids, RV_founder, founder_pat_allele, founder_mat_allele,
  haplos, RV_col_loc, RV_pool_loc)
}
\arguments{
\item{founder_ids}{Numeric list. The ID numbers of all founders.}

\item{RV_founder}{Numeric. The ID number(s) of the founder who introduced a cRV.}

\item{founder_pat_allele}{Numeric list. The paternally inherited allele for each founder.}

\item{founder_mat_allele}{Numeric list. The maternally inherited allele for each founder.}

\item{haplos}{sparseMatrix.  The sparseMatrix of genomes returned by \code{read_slimOut}.}

\item{RV_col_loc}{Numeric. The column location of the familial RV in haplos.}

\item{RV_pool_loc}{The column locations of each SNV (in haplos)  for SNVs in the pool of cRVs.}
}
\value{
list of familial founder genotypes
}
\description{
\strong{For internal use.}
}
\keyword{internal}
