% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_StudySeqFunctions.R
\name{sim_RVstudy}
\alias{sim_RVstudy}
\title{Simulate sequence data for a study}
\usage{
sim_RVstudy(ped_files, founder_genotypes, marker_map, chrom_map,
  affected_only = TRUE, convert_to_cM = TRUE, burn_in = 1000,
  gamma_params = c(2.63, 2.63/0.5))
}
\arguments{
\item{ped_files}{Data frame. Must match format of pedigree simulated by sim_RVped}

\item{founder_genotypes}{list of dataframes. Should contain, in the order listed in chrom_map, 1 dataframe for every chromosome contained in chrom_map.  Each data frame should contain two rows for each founder: the first to correspond to the paternally inherited gamete and the second to correspond to the maternally inherited gamete.}

\item{marker_map}{Data.frame. Must contain three columns with: column 1: marker names, must be listed in the same order as in the founder genotype file, column 2: the chromosomal position of the marker, column 3: the position of the marker in cM.}

\item{RV_markers}{character. A list of possible RV markers. If missing all markers in \code{marker_map} are assumed to be possible RV markers.}
}
\value{
study_sequences
}
\description{
Simulate sequence data for a study
}
\examples{
library(SimRVPedigree)
library(kinship2)
#Read in age-specific hazards
data(EgPeds)

EgPeds5 <- EgPeds4 <- EgPeds3 <- EgPeds2 <- EgPeds
EgPeds2$FamID <- EgPeds2$FamID + 5
EgPeds3$FamID <- EgPeds3$FamID + 10
EgPeds4$FamID <- EgPeds4$FamID + 15
EgPeds5$FamID <- EgPeds5$FamID + 20
ex_study_peds <- rbind(EgPeds, EgPeds2, EgPeds3, EgPeds4, EgPeds5)
ex_study_peds <- EgPeds

nrow(ex_study_peds[which(is.na(ex_study_peds$dad_id)), ])
length(unique(ex_study_peds$FamID))

my_chrom_map = data.frame(chrom     = c(1, 2),
                          start_pos = c(0, 0),
                          end_pos   = c(247199719, 242751149),
                          center = c(98879888, 97100460))
my_chrom_map

mark_map <- data.frame(chrom = c(1, 1, 1, 1, 1, 1, 1, 1,
                                      2, 2, 2, 2, 2, 2),
                       position = c(5012368, 5012369, 5012370,
                                    78541008, 78541009, 78541010,
                                    247199219, 247199220,
                                    11330, 11332,
                                    234577, 234578, 234579,
                                    18799180),
                       pathwayID = c(1, 1, 1, 2, 2, 2, 1, 1,
                                     2, 2, 1, 1, 1, 1),
                       possibleRV = c(0, 1, 1, 0, 0, 0, 1, 1,
                                      0, 0, 0, 1, 1, 1))
mark_map$marker <- paste0(mark_map$chrom, sep = "_", mark_map$position)
mark_map <- mark_map[, c(5, 1:4)]
mark_map

set.seed(1)
founder_seq <- as.data.frame(matrix(sample(2*nrow(mark_map)*1000,
                                    x = c(0, 1),
                                    replace = T,
                                    prob = c(0.95, 0.05)),
                             nrow = 2*1000))
colnames(founder_seq) = as.character(mark_map$marker)

set.seed(6)
ped_seq <- sim_RVstudy(ped_files = ex_study_peds,
                       founder_genotypes = founder_seq,
                       marker_map = mark_map,
                       chrom_map = my_chrom_map)
ped_seq

set.seed(6)
system.time(sim_RVstudy(ped_files = ex_study_peds,
                        founder_genotypes = founder_seq,
                        marker_map = mark_map,
                        chrom_map = my_chrom_map))

}

