% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_StudySeqFunctions.R
\name{sim_RVstudy}
\alias{sim_RVstudy}
\title{Simulate sequence data for a study}
\usage{
sim_RVstudy(ped_files, founder_genotypes, linkage_map, chrom_map, RV_markers,
  RV_marker_probs, affected_only = TRUE, burn_in = 1000,
  gamma_params = c(2.63, 2.63/0.5))
}
\arguments{
\item{ped_files}{Data frame. Must match format of pedigree simulated by sim_RVped}

\item{founder_genotypes}{list of dataframes. Should contain, in the order listed in chrom_map, 1 dataframe for every chromosome contained in chrom_map.  Each data frame should contain two rows for each founder: the first to correspond to the paternally inherited gamete and the second to correspond to the maternally inherited gamete.}

\item{linkage_map}{Data.frame. Must contain three columns with: column 1: marker names, must be listed in the same order as in the founder genotype file, column 2: the chromosomal position of the marker, column 3: the position of the marker in cM.}

\item{RV_markers}{character. A list of possible RV markers. If missing all markers in \code{linkage_map} are assumed to be possible RV markers.}
}
\value{
study_sequences
}
\description{
Simulate sequence data for a study
}
\examples{
library(SimRVPedigree)
library(kinship2)
#Read in age-specific hazards
data(EgPeds)

my_study_peds = EgPeds
length(my_study_peds[which(is.na(my_study_peds$dad_id)), ])
length(unique(my_study_peds$FamID))

my_chrom_map = data.frame(chrom     = c(1, 2),
                          start_pos = c(0, 0),
                          end_pos   = c(270, 200),
                          center = c(55, 40))
my_chrom_map

link_map <- data.frame(chrom = c(1, 1, 1, 1, 1, 1, 1,
                                      2, 2, 2, 2, 2, 2),
                       position = c(20, 50, 100, 125, 175, 200, 250,
                                     10, 25,  75, 125, 150, 200))
link_map$marker <- paste0(link_map$chrom, sep = "_", link_map$position)
link_map <- link_map[, c(3, 1, 2)]
link_map

my_RV_markers <- c('1_50', '2_75', '2_150')
my_RV_markers <- c('1_50')


set.seed(1)
founder_seq <- as.data.frame(matrix(sample(2*nrow(link_map)*1000,
                                    x = c(0, 1),
                                    replace = T,
                                    prob = c(0.95, 0.05)),
                             nrow = 2*1000))
colnames(founder_seq) = as.character(link_map$marker)
length(which(founder_seq$`1_50` == 1))


set.seed(6)
ped_seq <- sim_RVstudy(ped_files = my_study_peds,
                       founder_genotypes = founder_seq,
                       linkage_map = link_map,
                       chrom_map = my_chrom_map,
                       RV_markers = my_RV_markers)
ped_seq

set.seed(6)
system.time(sim_RVstudy(ped_files = my_study_peds,
                        founder_genotypes = founder_seq,
                        linkage_map = link_map,
                        chrom_map = my_chrom_map,
                        RV_marker = my_RV_marker))

}

