% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_StudySeqFunctions.R
\name{sim_RVstudy}
\alias{sim_RVstudy}
\title{Simulate sequence data for a sample of pedigrees}
\usage{
sim_RVstudy(ped_files, haplos, SNV_map, affected_only = TRUE,
  remove_wild = TRUE, pos_in_bp = TRUE, gamma_params = c(2.63, 2.63/0.5),
  burn_in = 1000)
}
\arguments{
\item{ped_files}{Data frame. A data frame of pedigrees for which to simulate sequence data, see details.}

\item{haplos}{sparseMatrix. A sparse matrix of haplotype data, which contains the haplotypes for unrelated individuals representing the founder population.  Rows are assumed to be haplotypes, while columns represent SNVs.  If the \code{\link{read_slim}} function was used to import SLiM data to \code{R}, users may supply the sparse matrix \code{Haplotypes} returned by \code{read_slim}.}

\item{SNV_map}{Data frame. A data frame that catalogs the SNVs in \code{haplos}.  If the \code{\link{read_slim}} function was used to import SLiM data to \code{R}, the data frame \code{Mutations} is of the proper format for \code{SNV_map}.  However, users must add the variable \code{is_CRV} to this data frame to identify causal rare variants, see details.}

\item{affected_only}{Logical. When \code{affected_only = TRUE}, we only simulate SNV data for the affected individuals and the family members that connect them along a line of descent.  When \code{affected_only = FALSE}, SNV data is simulated for the entire pedigree. By default, \code{affected_only = TRUE}.}

\item{remove_wild}{Logical.  A logical argument that determines if SNVs not carried by any member of the study should be removed from the data.  By default, \code{remove_wild = TRUE}.}

\item{pos_in_bp}{Logical. This argument indicates if the positions in \code{SNV_map} are listed in base pairs.  By default, \code{pos_in_bp = TRUE}. If the positions in \code{SNV_map} are listed in centiMorgan please set \code{pos_in_bp = FALSE} instead.}

\item{gamma_params}{Numeric list of length 2. The respective shape and rate parameters of the gamma distribution used to simulate distance between chiasmata.  By default, \code{gamma_params = c(2.63, 2*2.63)}, as discussed in Voorrips and Maliepaard (2012).}

\item{burn_in}{Numeric. The "burn-in" distance in centiMorgan, as defined by Voorrips and Maliepaard (2012), which is required before simulating the location of the first chiasmata with interference. By default, \code{burn_in = 1000}.
The burn in distance in cM. By default, \code{burn_in = 1000}.}
}
\value{
A object of class \code{famStudy}.  Objects of class \code{famStudy} are lists that will always include the following four items:

\item{\code{ped_files}}{A data frame containing the sample of pedigrees for which sequence data was simulated, see details.}

\item{\code{ped_haplos}}{A sparse matrix that contains the simulated haplotypes for each pedigree member in \code{ped_files}, see details.}

\item{\code{haplo_map}}{A data frame that maps the haplotypes (i.e. rows) in \code{ped_haplos} to the individuals in \code{ped_files}, see details.}

\item{\code{SNV_map}}{A data frame cataloging the SNVs in \code{ped_haplos}, see details.}
}
\description{
Simulate single-nucleotide variant (SNV) data for a sample of ascertained pedigrees.
}
\details{
Please note that \code{sim_RVstudy} was not created with the intention of simulating sequence data for pedigrees that contain inbreeding or loops.  Hence, \strong{certain type of inbreeding and/or loops can not be accomodated}. Please see examples.

The data frame of pedigrees, \code{ped_files}, supplied to \code{sim_RVstudy} must contain the variables:  \code{FamID}, \code{ID}, \code{sex}, \code{dadID}, \code{momID}, \code{affected}, \code{DA1}, and \code{DA2}. Please note: If the variables \code{DA1} and \code{DA2} are not provided we assume the pedigrees are fully sporadic.  The required variables are described as follows:
\tabular{lll}{
\strong{name} \tab \strong{type} \tab \strong{description} \cr
\code{FamID} \tab numeric \tab family identification number\cr
\code{ID} \tab numeric \tab individual identification number\cr
\code{sex} \tab numeric \tab sex identification variable: \code{sex = 0} for males, and \code{sex = 1} females. \cr
\code{dadID} \tab numeric \tab identification number of father \cr
\code{momID} \tab numeric \tab identification number of mother \cr
\code{affected} \tab logical \tab disease-affection status: \code{affected = TRUE} if individual has developed disease, and \code{FALSE} otherwise. \cr
\code{DA1} \tab numeric \tab paternally inherited allele at the cRV locus: \code{DA1 = 1} if the cRV is inherited, and \code{0} otherwise. \cr
\code{DA2} \tab numeric \tab maternally inherited allele at the cRV locus: \code{DA2 = 1} if the cRV is inherited, and \code{0} otherwise.\cr
}
The \code{\link{sim_RVped}} function of the \code{SimRVPedigree} package may be used to simulate pedigrees that meet this criteria.

The data frame \code{SNV_map}, which catalogs the SNVs in the argument \code{haplos}, must contain the following variables:
\tabular{lll}{
\strong{name} \tab \strong{type} \tab \strong{description} \cr
\code{colID} \tab numeric \tab associates the rows in \code{SNV_map} to the columns of \code{haplos}\cr
\code{chrom} \tab numeric \tab the chromosome the SNV resides on\cr
\code{position} \tab numeric \tab is the position of the SNV in base pairs when \code{pos_in_bp = TRUE}\cr
\code{afreq} \tab numeric \tab derived allele frequency of the SNV \cr
\code{marker} \tab character \tab a unique character identifier for the SNV \cr
\code{pathwaySNV} \tab logical \tab identifies SNVs located within the pathway of interest as \code{TRUE} \cr
\code{is_crv} \tab logical \tab  identifies causal rare variants (cRVs) as \code{TRUE}.  Note familial cRVs are sampled, with replacement from the SNVs for which \code{is_crv} is TRUE. \cr
}
Please note, the \code{read_slim} function format \code{Mutations} with all of the variables required for \code{SNV_map}, save \code{is_crv}, when importing SLIM data. Users are expected to create the variable \code{is_crv} prior to using the \code{sim_RVstudy} function. If the variable \code{is_crv} is not supplied, a single SNV will be chosen at random to be the causal rare variant for all pedigrees in the study.  Please refer to section 3.3 of the vignette for additional details regarding the variable \code{is_crv}.
}
\examples{
library(SimRVSequences)

#load pedigree, haplotype, and mutation data
data(study_peds)
data(EXmuts)
data(EXhaps)

#create variable is_CRV in EXmuts to identify causal
#rare variants, from which to sample familial variants.
EXmuts$is_CRV = FALSE
EXmuts$is_CRV[c(26, 73, 75, 238, 278)] = TRUE

#supply required inputs to the sim_RVstudy function
seqDat = sim_RVstudy(ped_files = study_peds,
                     SNV_map = EXmuts,
                     haplos = EXhaps)


# Inbreeding example
# Due to the forward-in-time model used by sim_RVstudy certain types of
# inbreeding and/or loops may cause fatal errors when using sim_RVstudy.
# The following two examples are used to demonstrate this.

# Create inbreeding in family 1 of study_peds
imb_ped1 <- study_peds[study_peds$FamID == 3, ]
imb_ped1[imb_ped1$ID == 18, c("momID")] = 7
plot(imb_ped1)

# Notice that this instance of inbreeding can be accomodated
# by the sim_RVstudy function
seqDat = sim_RVstudy(ped_files = imb_ped1,
                     SNV_map = EXmuts,
                     haplos = EXhaps)

# Create different type of inbreeding in family 1 of study_peds
imb_ped2 <- study_peds[study_peds$FamID == 3, ]
imb_ped2[imb_ped1$ID == 8, c("momID")] = 18
plot(imb_ped2)

# Notice that inbreeding in imb_ped2 will cause a fatal
# error when the sim_RVstudy function is executed
\dontrun{
seqDat = sim_RVstudy(ped_files = imb_ped2,
                     SNV_map = EXmuts,
                     haplos = EXhaps)
}

}
\references{
Roeland E. Voorrips and Chris A Maliepaard. (2012). \emph{The simulation of meiosis in diploid and tetraploid organisms using various genetic models}. BMC Bioinformatics, 13:248.

Christina Nieuwoudt and Jinko Graham. (??) Future bioRxiv article.
}
\seealso{
\code{\link{sim_RVped}}, \code{\link{read_slim}}, \code{\link{summary.famStudy}}
}
