% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_StudySeqFunctions.R
\name{sim_RVstudy}
\alias{sim_RVstudy}
\title{Simulate sequence data for a study}
\usage{
sim_RVstudy(ped_files, marker_map, chrom_map, haplotype_dist,
  affected_only = TRUE, convert_to_cM = TRUE, burn_in = 1000,
  gamma_params = c(2.63, 2.63/0.5))
}
\arguments{
\item{ped_files}{Data frame. Must match format of pedigree simulated by sim_RVped}

\item{marker_map}{Data.frame. Must contain three columns with: column 1: marker names, must be listed in the same order as in the founder genotype file, column 2: the chromosomal position of the marker, column 3: the position of the marker in cM.}

\item{chrom_map}{Data.frame.  A data.frame consisting of three columns: column 1 contains the chromosome numbers, column 2 start postion of chromosome (in cM), column 3 end position of chromosome (in cM).}

\item{burn_in}{Numeric. The burn in distance in cM.}

\item{gamma_params}{Numeric list of length 2. The respective shape and rate parameters gamma distribution used to simulate distance between chiasmata.  By default, \code{gamma_params = c(2.63, 2.63/0.5)}, as discussed in Vorrips (2012).}

\item{haplotype_dist.}{List of dataframes.  Each list entry should contain a dataframe containing the haplotype distribution for that chromosome (in same order as marker_map).}

\item{founder_genotypes}{list of dataframes. Should contain, in the order listed in chrom_map, 1 dataframe for every chromosome contained in chrom_map.  Each data frame should contain two rows for each founder: the first to correspond to the paternally inherited gamete and the second to correspond to the maternally inherited gamete.}

\item{RV_markers}{character. A list of possible RV markers. If missing all markers in \code{marker_map} are assumed to be possible RV markers.}
}
\value{
study_sequences
}
\description{
Simulate sequence data for a study
}
\examples{
library(SimRVPedigree)
data(EgPeds)

ex_study_peds <- EgPeds[EgPeds$FamID == 1, ]

data(hg_chrom)
my_chrom_map = hg_chrom
my_chrom_map

head(mark_map)
mm_obj <- markerMap(mark_map)
head(mm_obj)

h_dist <- list()
h_dist[[1]]<- SNP_dat

set.seed(6)
ped_seq <- sim_RVstudy(ped_files = ex_study_peds,
                       haplotype_dist = h_dist,
                       marker_map = mm_obj,
                       chrom_map = my_chrom_map)
ped_seq

}
