% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_StudySeqFunctions.R
\name{sim_RVstudy}
\alias{sim_RVstudy}
\title{Simulate sequence data for a study}
\usage{
sim_RVstudy(ped_files, marker_map, haplotype_dist, affected_only = TRUE,
  convert_to_cM = TRUE, burn_in = 1000, gamma_params = c(2.63, 2.63/0.5))
}
\arguments{
\item{ped_files}{Data frame. Must match format of pedigree simulated by sim_RVped}

\item{marker_map}{Data.frame. Must contain three columns with: column 1: marker names, must be listed in the same order as in the founder genotype file, column 2: the chromosomal position of the marker, column 3: the position of the marker in cM.}

\item{haplotype_dist}{sparseMatrix. The genomes matrix returned by \code{read_slim}.  Mutations in haplotype_dist are described in \code{marker_map}.}

\item{affected_only}{Logical. When \code{affected_only = TRUE} pedigrees are reduced contain only diesease-affected relatives, their parents, and any obligate carriers or founders; sequence data is simulated only for retained family memebres. When \code{affected_only = FALSE} sequence data is simulated for all.  By default, \code{affected_only = TRUE}.}

\item{convert_to_cM}{Logical. The setting \code{convert_to_cM = TRUE} must be used if genomic positions are given in base pairs.  If the genomic postions in \code{marker_map} are listed in centiMorgan, please set \code{convert_to_cM = FALSE}.  By default, \code{convert_to_cM = TRUE}.}

\item{burn_in}{Numeric. The burn in distance in cM. By default, \code{burn_in = 1000}.}

\item{gamma_params}{Numeric list of length 2. The respective shape and rate parameters gamma distribution used to simulate distance between chiasmata.  By default, \code{gamma_params = c(2.63, 2.63/0.5)}, as discussed in Vorrips (2012).}
}
\value{
study_sequences
}
\description{
Simulate sequence data for a study
}
\examples{
library(SimRVPedigree)
data(EgPeds)

library(SimRVSequences)
data(EXmut)
data(EXgen)

markers = EXmut
markers$possibleRV = FALSE
markers$possibleRV[1] = TRUE

seqDat = sim_RVstudy(ped_files = EgPeds,
                     marker_map = markers,
                     haplotype_dist = EXgen)

summary(seqDat)

}
