% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_StudySeqFunctions.R
\name{sim_RVstudy}
\alias{sim_RVstudy}
\title{Simulate sequence data for a study}
\usage{
sim_RVstudy(ped_files, SNV_map, haplos, affected_only = TRUE,
  pos_in_bp = TRUE, remove_wild = TRUE, burn_in = 1000,
  gamma_params = c(2.63, 2.63/0.5))
}
\arguments{
\item{ped_files}{Data frame. Must match format of pedigree simulated by sim_RVped}

\item{SNV_map}{Data.frame. Must contain three columns with: column 1: marker names, must be listed in the same order as in the founder genotype file, column 2: the chromosomal position of the marker, column 3: the position of the marker in cM.}

\item{haplos}{sparseMatrix. The genomes matrix returned by \code{read_slim}.  Mutations in haplos are described in \code{SNV_map}.}

\item{affected_only}{Logical. When \code{affected_only = TRUE} pedigrees are reduced contain only diesease-affected relatives, their parents, and any obligate carriers or founders; sequence data is simulated only for retained family memebres. When \code{affected_only = FALSE} sequence data is simulated for all.  By default, \code{affected_only = TRUE}.}

\item{pos_in_bp}{Logical. The setting \code{pos_in_bp = TRUE} must be used if genomic positions are given in base pairs.  If the genomic postions in \code{SNV_map} are listed in centiMorgan, please set \code{pos_in_bp = FALSE}.  By default, \code{pos_in_bp = TRUE}.}

\item{remove_wild}{Logical. Should markers at which no member of study carry a mutated allele be removed from the data. By default, \code{remove_wild = TRUE}.}

\item{burn_in}{Numeric. The burn in distance in cM. By default, \code{burn_in = 1000}.}

\item{gamma_params}{Numeric list of length 2. The respective shape and rate parameters gamma distribution used to simulate distance between chiasmata.  By default, \code{gamma_params = c(2.63, 2.63/0.5)}, as discussed in Vorrips (2012).}
}
\value{
study_sequences
}
\description{
Simulate sequence data for a study
}
\examples{
library(SimRVPedigree)
data(EgPeds)

library(SimRVSequences)
data(EXmuts)
data(EXgen)

markers = EXmuts
markers$possibleSNV = FALSE
markers$possibleSNV[1] = TRUE

seqDat = sim_RVstudy(ped_files = EgPeds,
                     SNV_map = markers,
                     haplos = EXgen)

summary(seqDat)

}
