% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVdata_methods.R
\name{SNVdata}
\alias{SNVdata}
\title{Constructor function for an object of class famStudy}
\usage{
SNVdata(SNV_data)
}
\arguments{
\item{SNV_data}{The list of items returned by the \code{read_slim} and \code{import_exons} function.}
}
\value{
an object of class \code{SNVdata}.
}
\description{
Constructor function for an object of class famStudy
}
