% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVdata_methods.R
\name{SNVdata}
\alias{SNVdata}
\title{Constructor function for an object of class SNVdata}
\usage{
SNVdata(Haplotypes, Mutations, Samples = NULL)
}
\arguments{
\item{Haplotypes}{sparseMatrix. A sparse matrix of haplotype data, which contains the haplotypes for unrelated individuals representing the founder population.  Rows are assumed to be haplotypes, while columns represent SNVs.}

\item{Mutations}{Data frame. A data frame that catalogs the SNVs in \code{Haplotypes}.}

\item{Samples}{An optional dataframe or matrix describing the individuals in \code{Haplotypes}.}
}
\value{
an object of class \code{SNVdata}.
}
\description{
Constructor function for an object of class SNVdata
}
