% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_seq_methods.R
\name{summary.famStudy}
\alias{summary.famStudy}
\title{Summary function for objects of class famStudy}
\usage{
\method{summary}{famStudy}(object, ...)
}
\arguments{
\item{object}{An object of class \code{famStudy}, returned by the \code{sim_RVstudy} function.}

\item{...}{additional arguments passed to other methods.}
}
\value{
\item{\code{fam_allele_count} }{A matrix that contains counts of the number of SNVs shared among the disease-affected relatives in each pedigree.}

\item{\code{pathway_count} }{A data frame that catalogs the total number of SNVs observed in study participants who are affected by disease.  When the variable \code{pathwaySNV} is included in the \code{SNV_map} data frame of the \code{famStudy} object, this data frame will identify SNVs in the pathway of interest as \code{TRUE}.}
}
\description{
Summary function for objects of class \code{famStudy}, that is objects returned by the \code{\link{sim_RVstudy}} function.
}
\examples{
library(SimRVSequences)

#load pedigree, haplotype, and mutation data
data(study_peds)
data(EXmuts)
data(EXhaps)

#create variable is_CRV in EXmuts to identify the causal
#rare variants from which to sample familial cRVs.
EXmuts$is_CRV = FALSE
EXmuts$is_CRV[c(26, 73, 75, 238, 278)] = TRUE

#supply required inputs to the sim_RVstudy function
seqDat = sim_RVstudy(ped_files = study_peds,
                     SNV_map = EXmuts,
                     haplos = EXhaps)

#to count the number of SNVs shared by the disease-affected
#relatives in each pedigree, supply the output returned by
#sim_RVstudy to the summary function
summary(seqDat)


}
\seealso{
\code{\link{sim_RVstudy}}
}
